+++ poo/src/main/java/com/poo/modelos/Profesor.java
@@ -4,13 +4,14 @@ import jakarta.persistence.Id;
 
 @Entity
 public class Profesor extends Persona {
+    
     @Id
     private String tipoContrato;
     
     public Profesor() {
     }
     
-    public Profesor(Double id, String nombres, String apellidos, String email, String tipoContrato) {
+    public Profesor(Integer id, String nombres, String apellidos, String email, String tipoContrato) {
         super(id, nombres, apellidos, email);
         this.tipoContrato = tipoContrato;
     }
@@ -26,8 +27,8 @@ public class Profesor extends Persona {
     @Override
     public String toString() {
         return "Profesor{" +
-               "tipoContrato='" + tipoContrato + '\'' +
-               ", " + super.toString() +
-               '}';
+            "tipoContrato='" + tipoContrato + '\'' +
+            ", " + super.toString() +
+            '}';
     }
 }
